Avoid redundant clipping path update when the clip-enabled state and the current clip path or region change in the same updateState() call.
If updateState() is called and both clip-enabled and clip-path have changed, the CG paint engine first
updated the clip path to no-clip or the current clip path and then later changed the clip path again
to the new path.

This commit only re-applies the current clip path or clears the clip if the clip path / region
have not changed in the same update.