Always use the Leopard version of the libWebKitSystemInterface library.
The Apple XCode config files provide a config option to set the deployment
version of OS X so that a WebKit build created on OS X 10.7 works on earlier
versions of OS X.

Qt always used the one for the Mac OS SDK that Qt was being built against.

This is a simple change to always use the Leopard version of the library.

See http://www.qtcentre.org/archive/index.php/t-47183.html for more details.