First part of fix for QTBUG-11401
When a Qt widget lost keyboard focus to a native Cocoa widget and the user
clicked on the Qt widget to restore focus, AppKit invoked [QCocoaView acceptsFirstResponder]
to check whether the widget can accept the focus but also then invokes [QCocoaView needsPanelToBecomeKey]
which was not implemented in QCocoaView.

Re-implement needsPanelToBecomeKey to always return true.

The documentation on how needsPanelToBecomeKey is used by AppKit is not very clear,
but based on the documentation that I can find, [NSWindow sendEvent] only makes an NSView
the first responder if it returns true for both methods.

A separate issue is that when alien widgets are enabled, [QCocoaView acceptsFirstResponder] returns
YES / NO based on the focusPolicy() of the top-level (native) widget, not the alien widget which
the user actually clicked on, so it may return NO even though the user clicked on a widget
which should accept focus.

Reviewed-by: Carles Pina