Fix problem returning focus to Qt-widget from native Cocoa views when alien widgets are enabled.
This is a further fix for QTBUG-11401.

When alien widgets are enabled and the user clicks in a focus-able alien widget,
the native parent QCocoaView which receives the acceptsFirstResponder request will
return YES/NO based on its own focus policy and not that of the alien widget
which the user actually clicked in, unless the native widget is a top-level window -
which is often but not always the case.

As a simple solution, if alien-widgets are enabled, always return YES from [QCocoaView acceptsFirstResponder].

For a more complete fix, we'd need to know the context of the [QCocoaView acceptsFirstResponder]
call, including the child widget which received the current mouse event.

Reviewed-by: Carles Pina