Allow the NSWindow delegate for Qt-created NSWindow instances to be replaced with a proxy delegate.
If an application wants to override the delegate's behavior for an NSWindow created by Qt
it can do this by creating a proxy delegate which handles certain functionality itself and forwards
other calls to the original delegate.

This is currently not possible because [NSWindow qt_qwidget] relies on the identity of the
delegate not being changed.

This commit changes [NSWindow qt_qwidget] to use a respondsToSelector: call on the delegate
to test whether it can return the QWidget associated with a window instead of testing
the identity of [NSWindow delegate].  This approach still works if the delegate has been replaced
by a proxy.