Fix crash when setting a graphics effect on a widget in Qt/Cocoa
This disables graphics effects for native widgets in Qt/Cocoa, as was
the case under Qt 4.7 and enables them for alien widgets.

For the alien-widget case, it sufficed to mark the widget has being
in a paint event prior to rendering the widget with the current
graphics effect.  For the non-alien case this results in a recursive-repaint
warning and a crash later on due to a QPainter not having a QPaintEngine set.
For the time being, this patch just disables graphics effects on non-alien widgets
under Qt/Cocoa.

QTBUG-23205
Reviewed-by: Carles Pina