Fix actions being disabled in top-level menus which were hidden when the parent window was initially shown
If a top-level menu was hidden when the associated window was shown, the items in that menu
remain disabled if the top-level menu is later shown.

For menus Qt creates itself, it uses manual-item enabling ([NSMenu setAutoenablesItems] is set to NO)
and [NSMenuItem setEnabled:] is used to enable/disable menu items.

The top level menu ([NSApp mainMenu]) is still using automatic item enabling however, so whether
an item gets enabled or not is determined by [QCocoaMenuLoader validateMenuItem:] - this function
just returns [NSMenuItem isEnabled] however.  Since [NSMenuItem setEnabled] has no effect when
automatic enabling is on, this always returns false if the menu is initially hidden.

This fix returns QAction::isEnabled() for top-level menu items when Cocoa asks whether it should
be enabled.

MD-18606
QTBUG-26399