Fix a bug where the keyboard modifiers for drag enter events were set incorrectly when the drag started in another application.
For drag events starting outside of the current application, Qt
was synthesizing the keyboard modifiers from the supported drag operations (!).
If the drag operations included Qt::LinkAction, this resulted in the modifiers
being set to Cmd+Alt, which later resulted in Qt::LinkAction being
set as the default drop action even though the user was not actually
pressing those keys.

This fixes the problem by using QApplication::queryKeyboardModifiers() to get the
actual state of the modifier keys if the drag event started outside the current
application.

QTBUG-26412
MD-18602