Fix QApplication::queryKeyboardModifiers() returning incorrect values when the application is in the background.
GetCurrentEventKeyModifiers() returns the modifiers from the last event which passed
through the current application, this doesn't get updated if the current application is not active.

When dragging an event from another (active) application over a Qt-based (inactive) application, this resulted
in the drag action not being updated if the user changed pressed keys whilst the mouse was over the Qt window.

This uses GetCurrentKeyModifiers() instead, which returns the actual state of the keyboard - better matching
the documentation for QApplication::queryKeyboardModifiers()

QTBUG-26413
MD-18602