Fix mouse button release not being detected after a drag operation completes.
QApplication::mouseButtons() got out of sync after a drag operation
completed as QApplicationPrivate::mouse_buttons was not updated to
reflect the mouse being released at the end of the drag operation and
no explicit mouse release event is received during drag-and-drop.

This issue combined with 4bc2550e657ef1029919a15448f6f045354eeb96
resulted in mouse moves with no button held down after the end
of a drag-and-drop operation acting as if the left mouse button
was held down.

The issue does not apply to Qt 5.3 where there is similar logic
to reset the internal var tracking the state of the mouse in
[QNSView draggedImage:]

QTBUG-27634
QTBUG-16635
MD-20112
MD-20113